# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Training Simulator API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  admin: 'admin',
  learner: 'learner',
  developer: 'developer',
} as const;

export interface UserProfile {
  userId: number;
  /** @nullable */
  id?: string | null;
  /** @nullable */
  email?: string | null;
  name: string;
  /** @nullable */
  username?: string | null;
  role: UserProfileRole;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  organizationId?: number | null;
  /** @nullable */
  authProvider?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface SyncUserInput {
  email: string;
  /** @nullable */
  name?: string | null;
  authProvider?: string;
}

export type OrganizationMembershipRole = typeof OrganizationMembershipRole[keyof typeof OrganizationMembershipRole];


export const OrganizationMembershipRole = {
  owner: 'owner',
  admin: 'admin',
  manager: 'manager',
  learner: 'learner',
} as const;

export type OrganizationMembershipStatus = typeof OrganizationMembershipStatus[keyof typeof OrganizationMembershipStatus];


export const OrganizationMembershipStatus = {
  active: 'active',
  inactive: 'inactive',
  invited: 'invited',
} as const;

export interface OrganizationMembership {
  membershipId: string;
  organizationId: number;
  organizationName: string;
  role: OrganizationMembershipRole;
  status: OrganizationMembershipStatus;
  createdAt: string;
}

export type OrgMemberRole = typeof OrgMemberRole[keyof typeof OrgMemberRole];


export const OrgMemberRole = {
  owner: 'owner',
  admin: 'admin',
  manager: 'manager',
  learner: 'learner',
} as const;

export type OrgMemberStatus = typeof OrgMemberStatus[keyof typeof OrgMemberStatus];


export const OrgMemberStatus = {
  active: 'active',
  inactive: 'inactive',
  invited: 'invited',
} as const;

export interface OrgMember {
  membershipId: string;
  userId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  role: OrgMemberRole;
  status: OrgMemberStatus;
  joinedAt: string;
  isActive: boolean;
}