# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Training Simulator API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User management and auth
  - name: organizations
    description: Organization and membership management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check (legacy)
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /health:
    get:
      operationId: getHealth
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found in local DB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/sync:
    post:
      operationId: syncUser
      tags: [users]
      summary: JIT-provision or update current Clerk user in local DB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUserInput"
      responses:
        "200":
          description: User synced (existing user updated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizations:
    get:
      operationId: getMyOrganizations
      tags: [organizations]
      summary: Get organizations the current user belongs to (with their role in each)
      responses:
        "200":
          description: List of organization memberships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizationMembership"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizations/{orgId}/members:
    get:
      operationId: getOrgMembers
      tags: [organizations]
      summary: Get all members of an organization (admin/owner/manager only)
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of organization members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgMember"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - not a member or insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        userId:
          type: number
        id:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        name:
          type: string
        username:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, learner, developer]
        team:
          type: ["string", "null"]
        organizationId:
          type: ["number", "null"]
        authProvider:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        createdAt:
          type: ["string", "null"]
      required:
        - userId
        - name
        - role
    SyncUserInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
        authProvider:
          type: string
      required:
        - email
    OrganizationMembership:
      type: object
      properties:
        membershipId:
          type: string
        organizationId:
          type: number
        organizationName:
          type: string
        role:
          type: string
          enum: [owner, admin, manager, learner]
        status:
          type: string
          enum: [active, inactive, invited]
        createdAt:
          type: string
      required:
        - membershipId
        - organizationId
        - organizationName
        - role
        - status
        - createdAt
    OrgMember:
      type: object
      properties:
        membershipId:
          type: string
        userId:
          type: number
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, manager, learner]
        status:
          type: string
          enum: [active, inactive, invited]
        joinedAt:
          type: string
        isActive:
          type: boolean
      required:
        - membershipId
        - userId
        - name
        - role
        - status
        - joinedAt
        - isActive