# lib/api-zod/src/generated/types/orgMember.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Training Simulator API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrgMemberRole } from './orgMemberRole';
import type { OrgMemberStatus } from './orgMemberStatus';

export interface OrgMember {
  membershipId: string;
  userId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  role: OrgMemberRole;
  status: OrgMemberStatus;
  joinedAt: string;
  isActive: boolean;
}