# lib/api-zod/src/generated/types/organizationMembership.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Training Simulator API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrganizationMembershipRole } from './organizationMembershipRole';
import type { OrganizationMembershipStatus } from './organizationMembershipStatus';

export interface OrganizationMembership {
  membershipId: string;
  organizationId: number;
  organizationName: string;
  role: OrganizationMembershipRole;
  status: OrganizationMembershipStatus;
  createdAt: string;
}