# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Training Simulator API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from './userProfileRole';

export interface UserProfile {
  userId: number;
  /** @nullable */
  id?: string | null;
  /** @nullable */
  email?: string | null;
  name: string;
  /** @nullable */
  username?: string | null;
  role: UserProfileRole;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  organizationId?: number | null;
  /** @nullable */
  authProvider?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  createdAt?: string | null;
}